서용자 정보 변경 에러 수정

+++ lib/services/api/api_auth.dart
@@ -254,8 +254,8 @@ class ApiAuth {
   }
 
   static Future<Response?> updateUserService({
-    required int height,
-    required int weight,
+    required double height,
+    required double weight,
     String? img,
   }) async {
     try {
